refactor: reorganize types into api/domain/ui structure
