+++ src/atoms/bookAtom.ts
@@ -1,6 +1,6 @@
 import { atom } from 'jotai';
 
-import { BookMode } from '@/types';
+import { BookMode } from '@/types/ui/book';
 
 export const bookModeAtom = atom<BookMode>('read');
 export const showQuotePageAtom = atom(true);
