+++ src/types/domain/book.ts
@@ -2,5 +2,3 @@ export type Quote = {
   page: number;
   quote: string;
 };
-
-export type BookMode = 'read' | 'edit';
