--- src/types/index.ts
@@ -1,3 +1,0 @@
-export * from './book/book';
-export * from './dropdown/dropdown';
-export * from './join/join';
