+++ src/types/ui/book.ts
@@ -1,0 +1,1 @@
+export type BookMode = 'read' | 'edit';
